requirements.txt
blinker==1.6.3
build==1.0.3
certifi==2025.4.26
click==8.1.7
elastic-transport==8.17.1
elasticsearch==9.0.1
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pip-tools==7.3.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
setuptools==80.1.0
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.0.1
wheel==0.41.3
